backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdf2image==1.17.0
Pillow>=10.4.0
pytesseract==0.3.10
google-cloud-vision==3.5.0
openai==1.54.3
httpx==0.27.2
pandas==2.1.4
openpyxl==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0